requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.1
dominate==2.9.1
install-jdk==1.1.0
PyYAML==6.0.2
setuptools==75.1.0
wheel==0.44.0
